requirements.txt
# Core dependencies
firebase-admin>=6.2.0
pytz>=2023.3
python-dotenv>=1.0.0
openai>=1.3.0
flask>=2.3.3
streamlit>=1.28.0
pandas>=2.1.0
plotly>=5.17.0
requests>=2.31.0
backoff>=2.2.1  # For retry logic 